feat: 카테고리 리스트 API 연결 및 컴포넌트 분리
